Fixed too early check of iteration limit.
Till now, the iteration limit was checked in the middle of the
iteration where it is still unknown whether the current basis is
optimal.  This leads to premature abort when we set the itartion limit
to the number of iterations needed by SoPlex.  Now we check the
iteration limit not in the SPxSolver::terminate() method, but at the
start of every iteration.