new system for warm start
The Precision Boosting algorithm needs to benefit from the warm start feature. To do that, last basis met in simplex are explicitely stored when quiting the solve. If we are testing feasibility the basis is stored in the specific dataarray _oldFeasBasisStatus(Rows/Cols), same for unboundedness in _oldUnbdBasisStatus(Rows/Cols). This allows to have simultaniously three basis for each certificate mode (standard LP, testing feasibility, testing unboundedness). Before starting the solve, we check if a basis is ready to be loaded, and if so load it.