mantissa precision levels: 168 (min for mpfr), 192, 284, etc
Previously the number of bits for the mantissa was 168, the minimum for boost::multiprecision::mpfr, then was multiplied by 3/2.
We do almost the same, but we change the second level to imitate QSopt_ex
To summup, we do first double precision (53 bits), then 168, 192 (instead of 252), then multiply by 3/2 so 284, 426 etc.
Another modification, the epsilon tolerance of the boosted solver has been udated to allow higher precision levels.