fix: use vectors instead of sets
While sets guarantees uniqueness in the entries, that may not be something the user wants. Furthermore, vectors are more usually more efficient when it comes to looping over them, which is the only operation performed in this case.
The user can still utilise a set and then convert it to a vector before passing it to the function.